Extract default GUI playback modules
